Move Repository To the Data folder.
